+++ 08_MVC/HaVan/BaiTapMVC/BaiTapMVC/Model/Calculator.swift
@@ -15,7 +15,7 @@ enum Operator {
     case devide
 }
 
-class Calculator {
+final class Calculator {
   
     func congHaiSo(_ firstNumber: Float, _ secondNumber: Float) -> Float {
         return firstNumber + secondNumber
@@ -43,24 +43,6 @@ class Calculator {
             return nhanHaiSo(firstNumber, secondNumber)
         case .devide:
             return chiaHaiSo(firstNumber, secondNumber)
-        }
-    }
-}
-
-class CalculatorFactory {
-    var num1: String?
-    var num2: String?
-    var op: Operator?
-    
-    func thucHienPhepTinh(phepTinh: String) -> Operator {
-        if phepTinh == "+" {
-            return .plus
-        } else if phepTinh == "-" {
-            return .minor
-        } else if phepTinh == "x" {
-            return .multiple
-        } else {
-            return .devide
         }
     }
 }
