+++ 08_MVC/HaVan/BaiTapMVC/BaiTapMVC/SceneDelegate.swift
@@ -21,4 +21,3 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window.makeKeyAndVisible()
     }
 }
-
